LayoutTests/http/tests/security/window-named-valueOf-expected.txt

CONSOLE MESSAGE: Threw exception: TypeError: Cannot set prototype of immutable prototype object
This passes if it doesn't alert the contents of innocent-victim.
